refactor: apply View/Container pattern to 3 components
- NavChats → NavChats (container) + NavChatsView (presentation)
  Container: data fetching, search state, group computation, persistence
  View: rendering search header, empty states, grouped conversation list

- LoginForm → LoginForm (container) + LoginFormView (presentation)
  Container: form state, API calls, navigation
  View: card layout, form fields, buttons, error display

- ConversationSidebarItem → container + ConversationSidebarItemView
  Container: route resolution (isSelected, href, age)
  View: EntityRow with status icon, title, menu content
