+++ frontend/components/nav-chats.tsx
@@ -1,21 +1,14 @@
 'use client';
 
-import { Calligraph } from 'calligraph';
-import { Inbox, Search } from 'lucide-react';
 import { useRouter } from 'next/navigation';
-import { Fragment, useEffect, useMemo, useState } from 'react';
-import { CollapsibleGroupHeader } from '@/components/collapsible-group-header';
-import { ConversationSearchHeader } from '@/components/conversation-search-header';
-import { ConversationSidebarItem } from '@/components/conversation-sidebar-item';
-import { ConversationsEmptyState } from '@/components/conversations-empty-state';
-import { SectionHeader } from '@/components/section-header';
+import { useEffect, useMemo, useState } from 'react';
+import { NavChatsView } from '@/components/nav-chats-view';
 import useGetConversations from '@/hooks/get-conversations';
 import {
   buildConversationGroups,
   countGroupItems,
   filterConversationGroups,
 } from '@/lib/conversation-groups';
-import { highlightMatch } from '@/lib/highlight-match';
 
 /** localStorage key used to persist which date groups the user has collapsed. */
 const COLLAPSED_GROUPS_STORAGE_KEY = 'nav-chats-collapsed-groups';
@@ -45,11 +38,11 @@ function loadCollapsedGroups(): Set<string> {
 }
 
 /**
- * Sidebar conversation list with search, date grouping, and collapsible sections.
+ * Container for the sidebar conversation list.
  *
- * Conversations are grouped by calendar day (newest-first). Groups with more
- * than one section support collapse/expand; collapsed state is persisted in
- * localStorage. A search bar filters by title once the query reaches 2+ chars.
+ * Owns data fetching (conversations), search state, group computation,
+ * collapsed-group persistence, and navigation. Delegates all rendering
+ * to `NavChatsView`.
  */
 export function NavChats(): React.JSX.Element {
   const router = useRouter();
@@ -99,90 +92,19 @@ export function NavChats(): React.JSX.Element {
     });
   };
 
-  // --- content resolution ---
-  // Computed outside JSX to avoid hard-to-read nested ternaries.
-  let content: React.JSX.Element | null = null;
-
-  if (isLoading) {
-    content = null;
-  } else if (!conversations?.length) {
-    content = (
-      <ConversationsEmptyState
-        icon={<Inbox className="h-4 w-4" />}
-        title="No sessions yet"
-        description="Sessions with your agent appear here. Start one to get going."
-        buttonLabel="New Session"
-        onAction={() => router.push('/')}
-      />
-    );
-  } else if (isSearchActive && resultCount === 0) {
-    content = (
-      <ConversationsEmptyState
-        icon={<Search className="h-4 w-4" />}
-        title="No matching sessions"
-        description="Try a different title fragment. Search lights up once you have at least two characters."
-      />
-    );
-  } else {
-    content = (
-      <div className="pt-1">
-        <ul className="flex w-full min-w-0 flex-col gap-0">
-          {filteredGroups.map((group) => {
-            // Only allow collapsing when there are multiple groups and
-            // the user is not searching (search always shows all matches).
-            const isCollapsible = !isSearchActive && filteredGroups.length > 1;
-
-            // Gate on isCollapsible so a persisted key can't hide items
-            // when only one group remains.
-            const isCollapsed = isCollapsible && collapsedGroups.has(group.key);
-
-            return (
-              <Fragment key={group.key}>
-                {isCollapsible ? (
-                  <CollapsibleGroupHeader
-                    label={group.label}
-                    isCollapsed={isCollapsed}
-                    itemCount={group.items.length}
-                    onToggle={() => toggleGroupCollapse(group.key)}
-                  />
-                ) : (
-                  <SectionHeader label={group.label} />
-                )}
-                {isCollapsed
-                  ? null
-                  : group.items.map((conversation, index) => (
-                      <ConversationSidebarItem
-                        key={conversation.id}
-                        id={conversation.id}
-                        title={
-                          isSearchActive ? (
-                            highlightMatch(conversation.title, searchQuery)
-                          ) : (
-                            <Calligraph>{conversation.title}</Calligraph>
-                          )
-                        }
-                        ariaLabel={conversation.title}
-                        updatedAt={conversation.updated_at}
-                        showSeparator={index > 0}
-                      />
-                    ))}
-              </Fragment>
-            );
-          })}
-        </ul>
-      </div>
-    );
-  }
-
   return (
-    <div className="flex min-h-0 flex-1 flex-col">
-      <ConversationSearchHeader
-        searchQuery={searchQuery}
-        onSearchChange={setSearchQuery}
-        onSearchClose={() => setSearchQuery('')}
-        resultCount={resultCount}
-      />
-      {content}
-    </div>
+    <NavChatsView
+      searchQuery={searchQuery}
+      onSearchChange={setSearchQuery}
+      onSearchClose={() => setSearchQuery('')}
+      resultCount={resultCount}
+      isLoading={isLoading}
+      isEmpty={!conversations?.length}
+      isSearchActive={isSearchActive}
+      filteredGroups={filteredGroups}
+      collapsedGroups={collapsedGroups}
+      onToggleGroup={toggleGroupCollapse}
+      onNewSession={() => router.push('/')}
+    />
   );
 }
